remove some code duplication in tests
