+++ tests/yadism/cf/test_nc_asy.py
@@ -1,39 +1,10 @@
 import numpy as np
-from eko.thresholds import ThresholdsAtlas
 
-from yadism import observable_name as on
 from yadism.coefficient_functions.fonll import f2_nc as f_f2_nc
 from yadism.coefficient_functions.heavy import f2_nc as h_f2_nc
 
 
-class MockCouplingConstants:
-    def get_weight(self, _pid, _q2, qct):
-        if qct == "VV":
-            return 1
-        if qct == "VA":
-            return 2
-        if qct == "AV":
-            return 4
-        if qct == "AA":
-            return 8
-        raise ValueError(f"Unkown {qct}")
-
-
-class MockSF:
-    def __init__(self, n):
-        self.obs_name = on.ObservableName(n)
-        self.coupling_constants = MockCouplingConstants()
-        self.m2hq = [1.0, 2.0, 3.0]
-        self.threshold = ThresholdsAtlas(self.m2hq)
-
-
-class MockESF:
-    def __init__(self, sf, x, Q2):
-        self.sf = MockSF(sf)
-        self.x = x
-        self.Q2 = Q2
-        self.process = "NC"
-
+from utils import MockESF
 
 def test_cg():
     Q2 = 200
